[DROOLS-7039] Move ResourceHelper to drools-util. Remove duplicates
