+++ drools-scenario-simulation/drools-scenario-simulation-backend/src/test/java/org/drools/scenariosimulation/backend/TestUtils.java
@@ -24,7 +24,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import org.drools.scenariosimulation.api.model.AuditLogLine;
-import org.drools.scenariosimulation.backend.util.ResourceHelper;
+import org.drools.util.ResourceHelper;
 import org.kie.api.builder.Message;
 import org.kie.dmn.api.core.DMNMessage;
