+++ drools-util/src/main/java/org/drools/util/ResourceHelper.java
@@ -14,10 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.efesto.common.api.utils;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+package org.drools.util;
 
 import java.io.File;
 import java.io.IOException;
@@ -31,18 +28,27 @@ import java.util.stream.Stream;
  */
 public class ResourceHelper {
 
-    private static final Logger logger = LoggerFactory.getLogger(ResourceHelper.class);
-
     /**
      * Scan into classpath folders to find resources with the required extension
      *
      * @param extension to find
      * @return stream of matching resources
      */
-    public static Stream<File> getResourcesByExtension(String extension) {
+    public static Stream<File> getFileResourcesByExtension(String extension) {
+        return Arrays.stream(getClassPathElements())
+                .flatMap(elem -> internalGetFileResources(elem, Pattern.compile(".*\\." + extension + "$")));
+    }
+
+    /**
+     * Scan into classpath folders to find resources with the required extension
+     * @param extension to find
+     * @return stream of matching resources
+     */
+    public static Stream<String> getResourcesByExtension(String extension) {
         return Arrays.stream(getClassPathElements())
                 .flatMap(elem -> internalGetResources(elem, Pattern.compile(".*\\." + extension + "$")));
     }
+
 
     /**
      * Scan folder to find resources that match with pattern
@@ -50,27 +56,53 @@ public class ResourceHelper {
      * @param directory where to start the search
      * @param pattern   to find
      * @return stream of matching resources
-     * @throws IOException
      */
-    public static Stream<File> getResourcesFromDirectory(File directory, Pattern pattern) {
+    public static Stream<File> getFileResourcesFromDirectory(File directory, Pattern pattern) {
         if (directory == null || directory.listFiles() == null) {
             return Stream.empty();
         }
         return Arrays.stream(Objects.requireNonNull(directory.listFiles())).flatMap(
                 elem -> {
                     if (elem.isDirectory()) {
-                        return getResourcesFromDirectory(elem, pattern);
+                        return getFileResourcesFromDirectory(elem, pattern);
                     } else {
                         try {
                             if (pattern.matcher(elem.getCanonicalPath()).matches()) {
                                 return Stream.of(elem);
                             }
                         } catch (final IOException e) {
-                            logger.error("Failed top retrieve resources from directory " + directory.getAbsolutePath() + " with pattern " + pattern.pattern(), e);
+                            throw new RuntimeException("Failed to retrieve resources from directory " + directory.getAbsolutePath() + " with pattern " + pattern.pattern(), e);
                         }
                     }
                     return Stream.empty();
                 });
+    }
+
+    /**
+     * Scan folder to find resources that match with pattern
+     * @param directory where to start the search
+     * @param pattern to find
+     * @return stream of matching resources
+     */
+    public static Stream<String> getResourcesFromDirectory(File directory, Pattern pattern) {
+        if (directory == null || directory.listFiles() == null) {
+            return Stream.empty();
+        }
+        return Arrays.stream(directory.listFiles()).flatMap(elem -> {
+            if (elem.isDirectory()) {
+                return getResourcesFromDirectory(elem, pattern);
+            } else {
+                try {
+                    String fileName = elem.getCanonicalPath();
+                    if (pattern.matcher(fileName).matches()) {
+                        return Stream.of(fileName);
+                    }
+                } catch (final IOException e) {
+                    throw new RuntimeException("Impossible to access to resources", e);
+                }
+            }
+            return Stream.empty();
+        });
     }
 
     static String[] getClassPathElements() {
@@ -84,7 +116,21 @@ public class ResourceHelper {
      * @param pattern to find
      * @return stream of matching resources
      */
-    static Stream<File> internalGetResources(String path, Pattern pattern) {
+    static Stream<File> internalGetFileResources(String path, Pattern pattern) {
+        final File file = new File(path);
+        if (!file.isDirectory()) {
+            return Stream.empty();
+        }
+        return getFileResourcesFromDirectory(file, pattern);
+    }
+
+    /**
+     * This method is internal because it works only with folder to explore (classPath folder) and not with exact paths
+     * @param path to folder or jar
+     * @param pattern to find
+     * @return stream of matching resources
+     */
+    static Stream<String> internalGetResources(String path, Pattern pattern) {
         final File file = new File(path);
         if (!file.isDirectory()) {
             return Stream.empty();
