+++ drools-util/src/test/java/org/drools/util/ResourceHelperTest.java
@@ -1,0 +1,144 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.drools.util;
+
+import java.io.File;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import org.junit.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.drools.util.ResourceHelper.getClassPathElements;
+import static org.drools.util.ResourceHelper.getFileResourcesByExtension;
+import static org.drools.util.ResourceHelper.getFileResourcesFromDirectory;
+import static org.drools.util.ResourceHelper.getResourcesByExtension;
+import static org.drools.util.ResourceHelper.getResourcesFromDirectory;
+import static org.drools.util.ResourceHelper.internalGetResources;
+
+public class ResourceHelperTest {
+
+    private static final String TEST_FILE = "TestFile.txt";
+
+    @Test
+    public void getResourcesByExtensionTest() {
+        Stream<String> txtResources = getResourcesByExtension("txt");
+        List<String> resources = txtResources.collect(Collectors.toList());
+        assertThat(resources.size()).isEqualTo(1);
+        assertThat(resources.stream().anyMatch(elem -> elem.endsWith(TEST_FILE))).isTrue();
+    }
+
+    @Test
+    public void getResourcesByExtensionExisting() {
+        final Stream<File> retrieved = getFileResourcesByExtension("txt");
+        commonVerifyStream(retrieved, TEST_FILE);
+    }
+
+    @Test
+    public void getResourcesByExtensionNotExisting() {
+        final Stream<File> retrieved = getFileResourcesByExtension("arg");
+        commonVerifyStream(retrieved, null);
+    }
+
+    @Test
+    public void getResourcesFromDirectoryTest() {
+        List<String> classPathElements = Arrays.asList(ResourceHelper.getClassPathElements());
+        Optional<String> testFolder = classPathElements.stream().filter(elem -> elem.contains("test-classes")).findFirst();
+        assertThat(testFolder.isPresent()).isTrue();
+        File dir = new File(testFolder.get());
+        String regex = ".*"+ TEST_FILE;
+        List<String> filesFound = getResourcesFromDirectory(dir, Pattern.compile(regex))
+                .collect(Collectors.toList());
+        assertThat(filesFound.size()).isEqualTo(1);
+
+        assertThat(getResourcesFromDirectory(null, null).count()).isEqualTo(0);
+        assertThat(getResourcesFromDirectory(dir, Pattern.compile("noMatch")).count()).isEqualTo(0);
+    }
+
+    @Test
+    public void getResourcesFromDirectoryExisting() {
+        File directory = new File("." + File.separator + "target" + File.separator + "test-classes");
+        Pattern pattern = Pattern.compile(".*txt");
+        final Stream<File> retrieved = getFileResourcesFromDirectory(directory, pattern);
+        commonVerifyStream(retrieved, TEST_FILE);
+    }
+
+    @Test
+    public void getResourcesFromDirectoryNotExisting() {
+        File directory = new File("." + File.separator + "target" + File.separator + "test-classes");
+        Pattern pattern = Pattern.compile(".*arg");
+        final Stream<File> retrieved = getFileResourcesFromDirectory(directory, pattern);
+        commonVerifyStream(retrieved, null);
+    }
+
+    @Test
+    public void getClassPathElements() {
+        String[] retrieved = ResourceHelper.getClassPathElements();
+        assertThat(retrieved).isNotNull();
+        assertThat(retrieved.length == 0).isFalse();
+    }
+
+    @Test
+    public void getClassPathElementsTest() {
+        List<String> classPathElements = Arrays.asList(ResourceHelper.getClassPathElements());
+        List<String> notJar = classPathElements.stream().filter(elem -> !elem.contains(".jar")).collect(Collectors.toList());
+        assertThat(notJar.stream().anyMatch(elem -> elem.contains("test-classes"))).isTrue();
+    }
+
+    @Test
+    public void internalGetResourcesTest() {
+        List<String> classPathElements = Arrays.asList(ResourceHelper.getClassPathElements());
+        Optional<String> testFolder = classPathElements.stream().filter(elem -> elem.contains("test-classes")).findFirst();
+        assertThat(testFolder.isPresent()).isTrue();
+        File dir = new File(testFolder.get());
+        List<String> filesFound = internalGetResources(testFolder.get(), Pattern.compile(".*\\.txt$")).collect(Collectors.toList());
+        assertThat(filesFound.size()).isEqualTo(1);
+
+        assertThat(internalGetResources(filesFound.get(0), Pattern.compile(".*\\.txt$")).count()).isEqualTo(0);
+    }
+
+    @Test
+    public void internalGetResourcesExisting() {
+        String path = "." + File.separator + "target" + File.separator + "test-classes";
+        Pattern pattern = Pattern.compile(".*txt");
+        final Stream<File> retrieved = ResourceHelper.internalGetFileResources(path, pattern);
+        commonVerifyStream(retrieved, TEST_FILE);
+    }
+
+    @Test
+    public void internalGetResourcesNotExisting() {
+        String path = "." + File.separator + "target" + File.separator + "test-classes";
+        Pattern pattern = Pattern.compile(".*arg");
+        final Stream<File> retrieved = ResourceHelper.internalGetFileResources(path, pattern);
+        commonVerifyStream(retrieved, null);
+    }
+
+    private void commonVerifyStream(final Stream<File> toVerify, String expectedFile) {
+        assertThat(toVerify).isNotNull();
+        final List<File> retrieved = toVerify.collect(Collectors.toList());
+        if (expectedFile != null) {
+            assertThat(retrieved).hasSize(1);
+            assertThat(retrieved).allMatch(file -> file.exists() && file.getName().equals(expectedFile));
+        } else {
+            assertThat(retrieved).isEmpty();
+        }
+    }
+}
