+++ efesto/efesto-core/efesto-common-api/src/main/java/org/kie/efesto/common/api/utils/FileUtils.java
@@ -33,6 +33,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import org.drools.util.ResourceHelper;
 import org.kie.efesto.common.api.exceptions.KieEfestoCommonException;
 import org.kie.efesto.common.api.io.MemoryFile;
 import org.slf4j.Logger;
@@ -57,7 +58,7 @@ public class FileUtils {
      */
     public static File getFile(String fileName) {
         String extension = fileName.substring(fileName.lastIndexOf('.') + 1);
-        File toReturn = ResourceHelper.getResourcesByExtension(extension)
+        File toReturn = ResourceHelper.getFileResourcesByExtension(extension)
                 .filter(file -> file.getName().equals(fileName))
                 .findFirst()
                 .orElse(null);
