+++ kie-test-util/pom.xml
@@ -103,6 +103,7 @@
       <groupId>org.apache.maven</groupId>
       <artifactId>maven-model</artifactId>
     </dependency>
+
   </dependencies>
 
 </project>
